pyproject.toml
[project]
name = "a2a-xrpl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.0.3",
    "langchain-openai>=1.0.1",
    "langgraph>=1.0.2",
    "python-dotenv>=1.2.1",
    "ipfshttpclient>=0.8.0a2",
    "xrpl-py>=2.0.0",
]
